Move colorsAdapter to a game folder
